feat(oop_polymorphism): add oop_polymorphism
Added oop_polymorphism.
Demonstrated basic concepts of polymorphism.
It happens when every derived class is diffenert operations,
but the function name is the same, so announcing the function name in
virtual type in based class.

Virtual function is forcing derived class to override function of based
class.
If it is not a override, it is just hiding function of based class.